# openapi_server/models/author.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Author(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, first_name=None, last_name=None):  # noqa: E501
        """Author - a model defined in OpenAPI

        :param id: The id of this Author.  # noqa: E501
        :type id: str
        :param first_name: The first_name of this Author.  # noqa: E501
        :type first_name: str
        :param last_name: The last_name of this Author.  # noqa: E501
        :type last_name: str
        """
        self.openapi_types = {
            'id': str,
            'first_name': str,
            'last_name': str
        }

        self.attribute_map = {
            'id': 'id',
            'first_name': 'FirstName',
            'last_name': 'LastName'
        }

        self._id = id
        self._first_name = first_name
        self._last_name = last_name

    @classmethod
    def from_dict(cls, dikt) -> 'Author':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Author of this Author.  # noqa: E501
        :rtype: Author
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Author.


        :return: The id of this Author.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Author.


        :param id: The id of this Author.
        :type id: str
        """

        self._id = id

    @property
    def first_name(self) -> str:
        """Gets the first_name of this Author.


        :return: The first_name of this Author.
        :rtype: str
        """
        return self._first_name

    @first_name.setter
    def first_name(self, first_name: str):
        """Sets the first_name of this Author.


        :param first_name: The first_name of this Author.
        :type first_name: str
        """

        self._first_name = first_name

    @property
    def last_name(self) -> str:
        """Gets the last_name of this Author.


        :return: The last_name of this Author.
        :rtype: str
        """
        return self._last_name

    @last_name.setter
    def last_name(self, last_name: str):
        """Sets the last_name of this Author.


        :param last_name: The last_name of this Author.
        :type last_name: str
        """

        self._last_name = last_name