# openapi_server/models/genre.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class Genre(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, name=None):  # noqa: E501
        """Genre - a model defined in OpenAPI

        :param id: The id of this Genre.  # noqa: E501
        :type id: str
        :param name: The name of this Genre.  # noqa: E501
        :type name: str
        """
        self.openapi_types = {
            'id': str,
            'name': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name'
        }

        self._id = id
        self._name = name

    @classmethod
    def from_dict(cls, dikt) -> 'Genre':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Genre of this Genre.  # noqa: E501
        :rtype: Genre
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Genre.


        :return: The id of this Genre.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Genre.


        :param id: The id of this Genre.
        :type id: str
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Genre.


        :return: The name of this Genre.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Genre.


        :param name: The name of this Genre.
        :type name: str
        """

        self._name = name